requirements.txt
absl-py==2.1.0
alabaster==0.7.16
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constants==0.6.0
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==4.2.11
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.14
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.8
imagesize==1.4.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.4.1
mpltern==1.0.4
mpmath==1.3.0
mtcnn==0.1.1
mutagen==1.47.0
namex==0.0.8
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
numpydoc==1.8.0
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
periodictable==2.0.1
pillow==10.3.0
platformdirs==4.3.3
plotly==6.0.1
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyrolite==0.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
redis==5.2.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rich-toolkit==0.14.8
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spotipy==2.24.0
SQLAlchemy==2.0.35
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
streamlit==1.38.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinydb==4.8.2
toml==0.10.2
tomli==2.2.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.14.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.35.0
virtualenv==20.26.6
watchdog==4.0.2
watchfiles==1.1.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
yt-dlp==2024.8.6
ytmusicapi==1.8.2
zipp==3.20.2
